Table view for shop page.
